updated the "make square" algorithm
I updated it to work based on bands, using two nested loops instead of one. This allows me to remove the "check if the next cell is populated" part, which should improve runtime (and makes the code easier to understand, in my opinion).

I explain the new approach in lines 31-40 (a big comment)

The  new algorithm appears to give the correct result.