Fixed formatting and redid (a little) the check_square method
the instructions on check_square are very confusing. I changed the function so that it now returns "False" if the proposed magic_square fails any of the tests. I also added a test that checks if the rows/columns/diagionals sum to the same thing. If they don't, then a lengthy error message is printed (and check_square returns false).

I also reformatted the main function (which was looking crazyyyy).

The code appears to run, I think that this will be the final version that I turn in.